Project/etl/requirements.txt
firebase-admin
pandas
python-dateutil
matplotlib
numpy
